+++ ChangeLog.md
@@ -2,6 +2,7 @@
 IMPROVEMENTS
 - add config option `configuration_backup_always` (default false), fix [1180](https://github.com/Altinity/clickhouse-backup/issues/1180)
 - add `clean_local_broken` CLI command and `POST /backup/clean/local_broken` REST API, fix [1178](https://github.com/Altinity/clickhouse-backup/issues/1178)
+- create pid file in `/var/lib/clickhouse/backup/backup_name/clickchouse-backup.pid` to avoid parallel work for the same backup even when `allow_parallel: true`, fix [966](https://github.com/Altinity/clickhouse-backup/issues/966)
 
 BUG FIXES
 - fix race condition for `server --watch` command, fix [1177](https://github.com/Altinity/clickhouse-backup/issues/1177)